Configure GitHub App
GitHub Apps require some configuration to be able to interact with the
GitHub platform. At a minimum, the app id and private key must be
provided so that the app can authenticate itself when making API calls.

Optionally, a webhook secret can be configured to cryptographically sign
the webhook payload and prevent tampering. While GitHub considers the
webhook secret an optional feature, octox enforces that it is set to
increase the security of its apps.