Call workflow for each event (#9)
octox now accepts a function that is called for every event that the
server receives. The function takes the event, and returns a result. If
the workflow succeeded, the result contains a serialize value that is
used as the body of the response. If the workflow failed, an error is
returned to the caller.