Initialize workflow objects using a constructor
The interface of octox has been refactored. It no longer takes an
initialized workflow object, but instead asks users to provide a
function that can create the object. The motivation for the change is
that GitHub Apps require some configuration, e.g. the app id and private
key, to communicate with GitHub's API. But these values are only
available after octox has been initialized itself.

An alternative approach would have been to expose the functions that
initialize the app from environment variables. But then users would have
to set the same settings twice, once for the workflow object and once
for octox.

The current implementation is considered a prototype, which can
hopefully be cleaned up a bit as we gather more knowledge about creating
workflows for octox.