Refactor error types (#20)
A first pass has been done to clean up and simplify the errors that the
crate exposes. The error handling is still not great, with a significant
mental overhead caused by the various different kinds of errors that the
crate can return.

As we build out the functionality of the crate and various different
workflows, we want to come back to this area and improve it further.